requirements.txt
azure-ai-ml==1.32.0
azure-identity==1.25.3
pyyaml==6.0.2
matplotlib==3.9.0
seaborn==0.13.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.8.0
xgboost==3.2.0
imbalanced-learn==0.14.1
mlflow==3.10.1
azureml-mlflow==1.60.0
jupyter==1.1.1
notebook==7.0.0